feat(bl-02): localização e tipo em Fazenda, Decimal em AreaRural, limite de fazendas
- Fazenda: tipo_propriedade, cep, logradouro, bairro, ibge_municipio_codigo, ccir, sigef_codigo
- AreaRural: area_hectares/area_hectares_manual Float → Numeric(12,4) (precisão fiscal)
- FazendaService: _check_limite_fazendas() — BusinessRuleError 422 se plano esgotado
- Migration 46646289cedf aplicada em PostgreSQL

+++ services/api/core/cadastros/propriedades/models.py
@@ -1,5 +1,6 @@
 import uuid
 from datetime import datetime, timezone
+from decimal import Decimal
 from sqlalchemy import String, Boolean, DateTime, Float, ForeignKey, Text, JSON, Numeric, Index
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from sqlalchemy import Uuid as UUID
@@ -85,12 +86,12 @@ class AreaRural(Base):
     descricao: Mapped[str | None] = mapped_column(Text, nullable=True)
 
     # Dimensões
-    area_hectares: Mapped[float | None] = mapped_column(
-        Float, nullable=True,
+    area_hectares: Mapped[Decimal | None] = mapped_column(
+        Numeric(12, 4), nullable=True,
         comment="Calculado do polígono ou preenchido manualmente"
     )
-    area_hectares_manual: Mapped[float | None] = mapped_column(
-        Float, nullable=True,
+    area_hectares_manual: Mapped[Decimal | None] = mapped_column(
+        Numeric(12, 4), nullable=True,
         comment="Override manual quando não há geometria"
     )
